create new models to handle league and team data
