+++ app/controllers/team/indexcontroller.php
@@ -1,21 +1,21 @@
 <?php
 namespace App\Controllers\Team;
 
-use App\Lib\Scraper\Fleaflicker;
-use App\Models;
+use App\Models\Team;
+use App\Models\Ranking;
 
 class IndexController
 {
     public function indexAction()
     {
         $id = $_GET["id"];
-        $team = new Fleaflicker\Team($id);
+        $team = new Team($id);
         $players = $team->getPlayers();
         $team_name = $team->getTeamName();
         $owner = $team->getOwner();
 
         for ($i = 0; $i < count($players); $i++) {
-            $ranking = new Models\Ranking($players[$i]["name"]);
+            $ranking = new Ranking($players[$i]["name"]);
             $players[$i]["rank"] = $ranking->getRank();
             $players[$i]["3week"] = $ranking->getDelta(3);
             $players[$i]["10week"] = $ranking->getDelta(10);
