+++ app/models/league.php
@@ -1,13 +1,15 @@
 <?php
-namespace App\Lib\Scraper\Fleaflicker;
+namespace App\Models;
 
 use App\Lib\Scraper\Scraper;
 use App\Lib\Config\Config;
+use App\Lib\Db\Models;
 
-class League extends Scraper
+class League extends Models
 {
     const XPATH_TEAM_DATA = "//td[contains(@class, 'left')][1]/div[contains(@class, 'league-name')]/a";
     const XPATH_OWNER_DATA = "//td[contains(@class,'right')][1]";
+    protected $scraper;
     protected $team_ids = [];
     protected $team_names = [];
     protected $team_owners = [];
@@ -17,8 +19,8 @@ class League extends Scraper
         $config = Config::getInstance();
         $league_type = $config["ff"]["type"];
         $league_id = $config["ff"]["league"];
-        $this->url = "https://www.fleaflicker.com/".$league_type."/leagues/".$league_id;
-        parent::__construct();
+        $url = "https://www.fleaflicker.com/".$league_type."/leagues/".$league_id;
+        $this->scraper = new Scraper($url);
     }
 
     public function getTeams()
@@ -42,7 +44,7 @@ class League extends Scraper
     protected function setTeamIds()
     {
         try {
-            $team_urls = $this->getHrefs(self::XPATH_TEAM_DATA);
+            $team_urls = $this->scraper->getHrefs(self::XPATH_TEAM_DATA);
         } catch (\Exception $e) {
             return;
         }
@@ -54,7 +56,7 @@ class League extends Scraper
     protected function setTeamNames()
     {
         try {
-            $this->team_names = $this->getTextContent(self::XPATH_TEAM_DATA);
+            $this->team_names = $this->scraper->getTextContent(self::XPATH_TEAM_DATA);
         } catch (\Exception $e) {
             return;
         }
@@ -63,7 +65,7 @@ class League extends Scraper
     protected function setTeamOwners()
     {
         try {
-            $this->team_owners = $this->getTextContent(self::XPATH_OWNER_DATA);
+            $this->team_owners = $this->scraper->getTextContent(self::XPATH_OWNER_DATA);
         } catch (\Exception $e) {
             return;
         }
