+++ batch/league_test.php
@@ -1,0 +1,7 @@
+<?php
+require("test_requirements.php");
+
+use App\Models\League;
+
+$league = new League();
+var_dump($league->getTeams());
