+++ batch/player_test.php
@@ -1,0 +1,8 @@
+<?php
+require("test_requirements.php");
+
+use App\Models\Player;
+
+$name = 'Tom Brady';
+$player = new Player($name);
+var_dump($player->getScore());
