+++ batch/team_test.php
@@ -1,0 +1,7 @@
+<?php
+require("test_requirements.php");
+
+use App\Models\Team;
+
+$team = new Team("1153300");
+var_dump($team->getPlayers());
