+++ batch/test_requirements.php
@@ -1,0 +1,5 @@
+<?php
+session_start();
+$_SESSION["env"] = "dev";
+$root = str_replace("/batch", "", __DIR__);
+require $root."/autoload.php";
